MainMenu MVC with saving and loading players and games
