+++ src/main/java/edu/ntnu/idi/idatt/boardgame/ui/javafx/view/MainMenu/MainMenuView.java
@@ -1,0 +1,280 @@
+package edu.ntnu.idi.idatt.boardgame.ui.javafx.view.MainMenu;
+
+import java.util.logging.Logger;
+import org.kordamp.ikonli.boxicons.BoxiconsLogos;
+import org.kordamp.ikonli.boxicons.BoxiconsRegular;
+import org.kordamp.ikonli.javafx.FontIcon;
+import edu.ntnu.idi.idatt.boardgame.Application;
+import edu.ntnu.idi.idatt.boardgame.Utils;
+import edu.ntnu.idi.idatt.boardgame.model.Player;
+import edu.ntnu.idi.idatt.boardgame.ui.javafx.IView;
+import javafx.collections.ListChangeListener;
+import javafx.collections.ObservableList;
+import javafx.geometry.Insets;
+import javafx.geometry.Pos;
+import javafx.scene.control.Button;
+import javafx.scene.control.ColorPicker;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.layout.BorderPane;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.Pane;
+import javafx.scene.layout.StackPane;
+import javafx.scene.layout.VBox;
+import javafx.scene.paint.Color;
+import javafx.scene.shape.Rectangle;
+import javafx.scene.text.Font;
+import javafx.scene.text.FontWeight;
+
+/**
+ * The main menu view of the application. This view displays available games and player management.
+ */
+public class MainMenuView implements IView {
+
+    private static final Logger logger = Logger.getLogger(MainMenuView.class.getName());
+
+    private BorderPane root;
+    private StackPane rootStack;
+    private MainMenuController controller;
+
+
+    private ListChangeListener<Player> playerSectionListener;
+
+    public MainMenuView(MainMenuController controller) {
+        this.controller = controller;
+    }
+
+    @Override
+    public void load() {
+        this.controller.initialize();
+        logger.info("Loading MainMenuView");
+    }
+
+    @Override
+    public void unload() {
+        logger.info("Unloading MainMenuView");
+        if (this.playerSectionListener != null) {
+            this.controller.getPlayers().removeListener(this.playerSectionListener);
+        }
+    }
+
+    @Override
+    public Pane createRoot() {
+        // Create the root layout with a StackPane as main container for modals/dialogs
+        rootStack = new StackPane();
+        root = new BorderPane();
+        root.setPadding(new Insets(10));
+
+        // Create the title header
+        var titleLabel = new Label("Awesome Board Games");
+        titleLabel.setFont(Font.font("System", FontWeight.BOLD, 40));
+        BorderPane.setAlignment(titleLabel, Pos.CENTER);
+        root.setTop(titleLabel);
+
+        // Create players section on the left
+        var playersSection = createPlayersSection();
+        root.setLeft(playersSection);
+
+        var rightSection = new VBox(20);
+        rightSection.setPrefWidth(300);
+        root.setRight(rightSection);
+
+        // Create games section on the right
+        var menuButtons = new VBox(20);
+        Application.getGameManager().getAvailableGameIds().forEach(gameId -> {
+            var gameButton = new Button(Application.getGameManager().getGame(gameId).getName());
+            gameButton.setPrefWidth(400);
+            gameButton.setPrefHeight(60);
+            gameButton.setOnAction(e -> this.controller.startGame(gameId));
+            menuButtons.getChildren().add(gameButton);
+        });
+
+        var exitButton = new Button("Exit to Desktop");
+        exitButton.setPrefWidth(400);
+        exitButton.setPrefHeight(60);
+        exitButton.setOnAction(e -> this.controller.exitToDesktop());
+        menuButtons.getChildren().add(exitButton);
+
+        menuButtons.setAlignment(Pos.CENTER);
+
+        root.setCenter(menuButtons);
+
+        // Add the main layout to the stack
+        rootStack.getChildren().add(root);
+
+        return rootStack;
+    }
+
+    private VBox createPlayersSection() {
+        VBox playersSection = new VBox(20);
+        playersSection.setPadding(new Insets(10));
+        playersSection.setPrefWidth(400);
+
+        Label playersLabel = new Label("Players");
+        playersLabel.setFont(Font.font("System", FontWeight.BOLD, 24));
+
+
+        // Create player entries in a scrollable container
+        VBox playerEntries = new VBox(10);
+        ScrollPane scrollPane = new ScrollPane(playerEntries);
+        scrollPane.setFitToWidth(true);
+        scrollPane.setPrefHeight(300);
+        scrollPane.setStyle("-fx-background: transparent; -fx-background-color: transparent;");
+
+        if (this.playerSectionListener != null) {
+            this.controller.getPlayers().removeListener(this.playerSectionListener);
+        }
+
+        this.playerSectionListener = e -> {
+            createPlayerEntries(playerEntries, this.controller.getPlayers());
+        };
+        this.controller.getPlayers().addListener(playerSectionListener);
+
+        createPlayerEntries(playerEntries, this.controller.getPlayers());
+
+        // Create "Add New Player" button and section
+        Button addPlayerButton = new Button("Add New Player");
+        addPlayerButton.setPrefWidth(380);
+        addPlayerButton.setStyle("-fx-background-color: #4CAF50; -fx-text-fill: white;");
+        addPlayerButton.setOnAction(e -> showAddPlayerPane(playerEntries));
+
+        // Create Save Players button
+        Button savePlayersButton = new Button("Save Players");
+        savePlayersButton.setPrefWidth(380);
+        savePlayersButton.setStyle("-fx-background-color: #2196F3; -fx-text-fill: white;");
+        savePlayersButton.setOnAction(e -> this.controller.savePlayersToFile());
+
+        playersSection.getChildren().addAll(playersLabel, scrollPane, addPlayerButton,
+                savePlayersButton);
+        return playersSection;
+    }
+
+    private void createPlayerEntries(VBox container, ObservableList<Player> players) {
+        container.getChildren().clear();
+
+        players.forEach(player -> {
+            HBox playerEntry = new HBox(10);
+
+            playerEntry.setAlignment(Pos.CENTER_LEFT);
+            playerEntry.setPadding(new Insets(5));
+            playerEntry.setStyle("-fx-background-color: #333333; -fx-background-radius: 5;");
+
+            // Player name field
+            TextField playerNameField = new TextField(player.getName());
+            playerNameField.setPrefWidth(120);
+            playerNameField.setStyle("-fx-background-color: #555555; -fx-text-fill: white;");
+
+            ColorPicker colorPicker = new ColorPicker(Utils.toJFXColor(player.getColor()));
+            // Remove player button
+            Button saveButton = new Button();
+            var saveIcon = new FontIcon(BoxiconsRegular.SAVE);
+            saveIcon.setFill(Color.WHITE);
+            saveButton.setGraphic(saveIcon);
+            saveButton.setStyle(
+                    "-fx-background-color: #2ecc71; -fx-text-fill: white; -fx-background-radius: 15; -fx-min-width: 30px; -fx-min-height: 30px;");
+            saveButton.setOnAction(e -> {
+                this.controller.updatePlayer(player, playerNameField.getText(),
+                        Utils.toModelColor(colorPicker.getValue()));
+            });
+            // Remove player button
+            Button removeButton = new Button();
+            var removeIcon = new FontIcon(BoxiconsRegular.TRASH);
+            removeIcon.setFill(Color.WHITE);
+            removeButton.setGraphic(removeIcon);
+            removeButton.setStyle(
+                    "-fx-background-color: #e74c3c; -fx-fill: white; -fx-background-radius: 15; -fx-min-width: 30px; -fx-min-height: 30px;");
+            removeButton.setOnAction(e -> {
+                this.controller.removePlayer(player);
+            });
+
+            // Add all components to the player entry
+            playerEntry.getChildren().addAll(playerNameField, colorPicker, saveButton,
+                    removeButton);
+            container.getChildren().add(playerEntry);
+        });
+    }
+
+
+    private void showAddPlayerPane(VBox playerEntries) {
+        // Create a semi-transparent dark overlay
+        Rectangle overlay = new Rectangle();
+        overlay.setWidth(2000); // Make it large enough to cover the entire view
+        overlay.setHeight(2000);
+        overlay.setStyle("-fx-background-color: rgba(0, 0, 0, 0.7);");
+
+        // Create a form to add a new player
+        VBox addPlayerForm = new VBox(10);
+        addPlayerForm.setPadding(new Insets(20));
+        addPlayerForm.setMaxWidth(400);
+        addPlayerForm.setMaxHeight(500);
+        addPlayerForm.setStyle(
+                "-fx-background-color: #333333; -fx-border-color: #666666; -fx-border-width: 2px; -fx-background-radius: 5;");
+
+        // Center the form in the stack pane
+        StackPane.setAlignment(addPlayerForm, Pos.CENTER);
+
+        Label formTitle = new Label("Add New Player");
+        formTitle.setFont(Font.font("System", FontWeight.BOLD, 16));
+        formTitle.setStyle("-fx-text-fill: white;");
+
+        TextField nameField = new TextField();
+        nameField.setPromptText("Player Name");
+        nameField.setPrefWidth(360);
+        nameField.setStyle("-fx-background-color: #555555; -fx-text-fill: white;");
+
+        Label nameLabel = new Label("Player Name:");
+        nameLabel.setStyle("-fx-text-fill: white;");
+
+        // Color selection
+        Label colorLabel = new Label("Player Color:");
+        colorLabel.setStyle("-fx-text-fill: white;");
+
+        ColorPicker colorPicker = new ColorPicker(javafx.scene.paint.Color.LIGHTBLUE);
+        colorPicker.setPrefWidth(360);
+
+        // Icon selection
+        Label iconLabel = new Label("Player Icon:");
+        iconLabel.setStyle("-fx-text-fill: white;");
+
+        Label previewLabel = new Label("Preview:");
+        previewLabel.setStyle("-fx-text-fill: white;");
+
+        HBox previewBox = new HBox(10);
+        previewBox.setAlignment(Pos.CENTER);
+        previewBox.getChildren().addAll(previewLabel);
+
+        // Add player button
+        Button addButton = new Button("Add Player");
+        addButton.setPrefWidth(360);
+        addButton.setStyle("-fx-background-color: #4CAF50; -fx-text-fill: white;");
+        addButton.setOnAction(e -> {
+            if (nameField.getText().trim().isEmpty()) {
+                return; // Don't add player with empty name
+            }
+
+            var newPlayer =
+                    new Player(nameField.getText(), Utils.toModelColor(colorPicker.getValue()));
+            this.controller.addPlayer(newPlayer);
+
+            // Remove both the form and overlay
+            rootStack.getChildren().removeAll(overlay, addPlayerForm);
+        });
+
+        // Cancel button
+        Button cancelButton = new Button("Cancel");
+        cancelButton.setPrefWidth(360);
+        cancelButton.setStyle("-fx-background-color: #FF5555; -fx-text-fill: white;");
+        cancelButton.setOnAction(e -> {
+            rootStack.getChildren().removeAll(overlay, addPlayerForm);
+        });
+
+        addPlayerForm.getChildren().addAll(formTitle, nameLabel, nameField, colorLabel, colorPicker,
+                iconLabel, previewBox, new VBox(10), // Spacer
+                addButton, cancelButton);
+
+        // Add both the overlay and form to the stack pane
+        // Adding the overlay first ensures it appears behind the form
+        rootStack.getChildren().addAll(overlay, addPlayerForm);
+    }
+}
